+++ Data/DataFunctions/Properties/AssemblyInfo.cs
@@ -19,5 +19,5 @@ using System.Reflection;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("1.9.41")]
-[assembly: AssemblyFileVersion("1.9.41")]
+[assembly: AssemblyVersion("1.10.6")]
+[assembly: AssemblyFileVersion("1.10.6")]