+++ Engine/Common/AvatarItem.cs
@@ -1,4 +1,5 @@
 ﻿using JocysCom.ClassLibrary.ComponentModel;
+using JocysCom.VS.AiCompanion.Plugins.Core.TtsMonitor;
 using System;
 using System.ComponentModel;
 using System.Linq;
@@ -14,6 +15,10 @@ namespace JocysCom.VS.AiCompanion.Engine
 			JocysCom.ClassLibrary.Runtime.Attributes.ResetPropertiesToDefault(this);
 		}
 
+		[DefaultValue(false)]
+		public bool AlwaysOnTop { get => _AlwaysOnTop; set => SetProperty(ref _AlwaysOnTop, value); }
+		private bool _AlwaysOnTop;
+
 		public Guid AiServiceId { get => _AiServiceId; set => SetProperty(ref _AiServiceId, value); }
 		Guid _AiServiceId;
 
@@ -28,18 +33,21 @@ namespace JocysCom.VS.AiCompanion.Engine
 		public string VoiceName { get => _VoiceName; set => SetProperty(ref _VoiceName, value); }
 		string _VoiceName;
 
-
 		[DefaultValue(null)]
-		public string VoiceLanguage { get => _VoiceLanguage; set => SetProperty(ref _VoiceLanguage, value); }
-		string _VoiceLanguage;
+		public string VoiceLocale { get => _VoiceLocale; set => SetProperty(ref _VoiceLocale, value); }
+		string _VoiceLocale;
 
+		[DefaultValue(VoiceGender.Male)]
+		public VoiceGender Gender { get => _Gender; set => SetProperty(ref _Gender, value); }
+		VoiceGender _Gender;
+
+		/// <summary>
+		/// Instructions to send when avatar is visible.
+		/// </summary>
 		[DefaultValue(null)]
-		public BindingList<string> VoiceNames
-		{
-			get => _VoiceNames = _VoiceNames ?? new BindingList<string>() { VoiceName };
-			set => SetProperty(ref _VoiceNames, value);
-		}
-		BindingList<string> _VoiceNames;
+		public string Instructions { get => _Instructions; set => SetProperty(ref _Instructions, value); }
+		string _Instructions;
+
 
 		[XmlIgnore, JsonIgnore]
 		public string AiServiceName { get => Global.AppSettings?.AiServices?.FirstOrDefault(x => x.Id == AiServiceId)?.Name; }